# backend/docs/swagger.yaml
basePath: /
definitions:
...
  entity.CreateIssueLinkReq:
    properties:
      link_type:
        enum:
        - relates_to
        - blocks
        - blocked_by
        - duplicates
        - is_duplicated_by
        type: string
      target_id:
        type: string
    required:
    - link_type
    - target_id
    type: object
...
  entity.CreateLabelReq:
    properties:
      color:
        type: string
      name:
        maxLength: 64
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  entity.File:
    properties:
      created_at:
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      id:
        type: string
      mime_type:
        type: string
      storage_path:
        type: string
      uploaded_by:
        type: string
      url:
        description: presigned download URL
        type: string
    type: object
...
  entity.Issue:
    properties:
      assignee:
        $ref: '#/definitions/entity.UserShort'
      assignee_id:
        type: string
      created_at:
        type: string
      custom_fields:
        additionalProperties: {}
        type: object
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      issue_number:
        type: integer
      labels:
        items:
          $ref: '#/definitions/entity.Label'
        type: array
      parent_id:
        type: string
      priority:
        description: lowest | low | medium | high | highest
        type: string
      project_id:
        type: string
      reporter:
        $ref: '#/definitions/entity.UserShort'
      reporter_id:
        type: string
      sprint_id:
        type: string
      status:
        $ref: '#/definitions/entity.WorkflowStatus'
      status_id:
        type: string
      story_points:
        type: integer
      title:
        type: string
      type:
        description: task | bug | story | epic | subtask
        type: string
      updated_at:
        type: string
    type: object
...
  entity.IssueLink:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      creator:
        $ref: '#/definitions/entity.UserShort'
      id:
        type: string
      link_type:
        description: relates_to | blocks | blocked_by | duplicates | is_duplicated_by
        type: string
      source:
        $ref: '#/definitions/entity.Issue'
      source_id:
        type: string
      target:
        $ref: '#/definitions/entity.Issue'
      target_id:
        type: string
    type: object
...
  entity.Label:
    properties:
      color:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
    type: object
...
  entity.UpdateLabelReq:
    properties:
      color:
        type: string
      name:
        maxLength: 64
        minLength: 1
        type: string
    type: object
...
  entity.UserShort:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
    type: object
...
  entity.WorkflowStatus:
    properties:
      category:
        description: todo | in_progress | done
        type: string
      color:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_initial:
        type: boolean
      name:
        type: string
      position:
        type: integer
      updated_at:
        type: string
      workflow_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Project management backend API (Jira-like).
  title: JiraFlow API
  version: "1.0"
paths:
...
  /api/v1/files/presign:
    get:
      parameters:
      - description: Storage path returned from upload
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  url:
                    type: string
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get presigned URL for a stored file
      tags:
      - files
  /api/v1/files/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/entity.File'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload a standalone file to object storage
      tags:
      - files
...
  /api/v1/issues/{issue_id}/links:
    get:
      parameters:
      - description: Issue ID
        in: path
        name: issue_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/entity.IssueLink'
                type: array
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List links for an issue
      tags:
      - issue-links
    post:
      consumes:
      - application/json
      parameters:
      - description: Source Issue ID
        in: path
        name: issue_id
        required: true
        type: string
      - description: Link data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreateIssueLinkReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/entity.IssueLink'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Link two issues
      tags:
      - issue-links
...
  /api/v1/issues/links/{id}:
    delete:
      parameters:
      - description: Link ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete issue link
      tags:
      - issue-links
  /api/v1/labels/{id}:
    delete:
      parameters:
      - description: Label ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete label
      tags:
      - labels
    get:
      parameters:
      - description: Label ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/entity.Label'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get label by ID
      tags:
      - labels
    put:
      consumes:
      - application/json
      parameters:
      - description: Label ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateLabelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/entity.Label'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update label
      tags:
      - labels
...
  /api/v1/projects/{project_id}/labels:
    get:
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/entity.Label'
                type: array
            type: object
      security:
      - BearerAuth: []
      summary: List labels by project
      tags:
      - labels
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Label data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreateLabelReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/entity.Label'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create label
      tags:
      - labels
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"